refactor: reorganise data transfer objects
Reorganised DTOs into input and output namespaces
