+++ app/Application/Actions/CreateUserAction.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Application\Actions;
 
-use App\Application\DTOs\CreateUserDTO;
-use App\Application\DTOs\UserResponseDTO;
+use App\Application\DTOs\Input\CreateUserDTO;
+use App\Application\DTOs\Output\UserResponseDTO;
 use InvalidArgumentException;
 use JsonException;
 use Psr\Http\Message\ResponseInterface as Response;
