+++ app/Application/Actions/ShowUserAction.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Application\Actions;
 
-use App\Application\DTOs\UserResponseDTO;
+use App\Application\DTOs\Output\UserResponseDTO;
 use App\Application\Exceptions\UserNotFoundException;
 use JsonException;
 use Psr\Http\Message\ResponseInterface as Response;
