+++ app/Application/Actions/UpdateUserAction.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Application\Actions;
 
-use App\Application\DTOs\UpdateUserDTO;
-use App\Application\DTOs\UserResponseDTO;
+use App\Application\DTOs\Input\UpdateUserDTO;
+use App\Application\DTOs\Output\UserResponseDTO;
 use App\Application\Exceptions\UserNotFoundException;
 use JsonException;
 use Psr\Http\Message\ResponseInterface as Response;
